DRY MarginVisualizer + PaddingVisualizer -> SpacingVisualizer
